### 2021-02-02 #425 @wlandau
jagstargets: reproducible JAGS pipelines at scale
Submitting Author: Will Landau (@wlandau)
Repository: https://github.com/wlandau/jagstargets
Version submitted:   0.0.0.9000
Version accepted: TBD

---

```
Package: jagstargets
Title: Targets for JAGS Workflows
Description: Bayesian data analysis usually incurs long runtimes
  and cumbersome custom code. A pipeline toolkit tailored to Bayesians,
  the 'jagstargets' R package is leverages
  'targets' and 'R2jags' to ease this burden.
  'jagstargets' makes it super easy to set up useful scalable
  JAGS pipelines that automatically parallelize the computation
  and skip expensive steps when the results are already up to date.
  Minimal custom code is required, and there is no need to manually
  configure branching, so usage is much easier than 'targets' alone.
  For the underlying methodology, please refer
  to the documentation of 'targets' <doi:10.21105/joss.02959> and 'JAGS'
  (Plummer 2003) <https://www.r-project.org/conferences/DSC-2003/Proceedings/Plummer.pdf>.
Version: 0.0.0.9000
  person(
    given = c("William", "Michael"),
    family = "Landau",
    email = "will.landau@gmail.com",
    comment = c(ORCID = "0000-0003-1878-3253")
  ),
  person(
    family = "Eli Lilly and Company",
    role = "cph"
  ))
  coda (>= 0.19.4),
  digest (>= 0.6.27),
  fst (>= 0.9.4),
  posterior (>= 0.1.3),
  purrr (>= 0.3.4),
  qs (>= 0.14.1),
  R2jags (>= 0.6.1),
  rjags (>= 4.10),
  rlang (>= 0.4.8),
  targets (>= 0.0.1),
  tarchetypes (>= 0.0.1),
  tibble (>= 3.0.4),
  tools,
  withr (>= 2.3.0),
  fs (>= 1.5.0),
  knitr (>= 1.28),
  R.utils (>= 2.10.1),
  rmarkdown (>= 2.1),
  tidyr (>= 1.1.2),
  visNetwork (>= 2.0.9)
  stan-dev/posterior
SystemRequirements: JAGS 4.x.y (http://mcmc-jags.sourceforge.net)
```

`jagstargets` leverages the existing workflow automation capabilities of [`targets`](https://docs.ropensci.org/targets) to orchestrate computation and skip up-to-date tasks in Bayesian data analysis pipelines. `jagstargets` reduces the burden of user-side custom code that [`targets`](https://docs.ropensci.org/targets) would otherwise require, which helps free statisticians to focus more on the models and less on the software engineering.

`jagstargets` is for Bayesian statisticians who develop and run JAGS models. Example workflows range from individual analyses of clinical data to large-scale simulation-based calibration studies for validation.

[`targets`](https://docs.ropensci.org/targets) already provides the same kind of workflow automation, but it requires more custom code to set up a workflow. `jagstargets` uses specialized domain knowledge to make this process easier. Packages `rjags` and `R2jags` interface with JAGS but provide little workflow automation.

N/A

N/A

~~The continuous integration is temporarily having some issues because `tarchetypes` was just accepted to CRAN and the binaries are not yet available.~~ (Should be fixed now.)

__This package depends on [`posterior`](https://github.com/stan-dev/posterior), which is not yet on CRAN.__

__I have prepared a [JOSS manuscript](https://github.com/wlandau/jagstargets/blob/main/inst/paper.md), and I intend to submit it if `jagstargets` clears rOpenSci software review.__
