+++ .github/workflows/pkgdown.yaml
@@ -38,7 +38,7 @@ jobs:
       - name: Install dependencies
         run: |
           install.packages("remotes")
-          remotes::install_deps(dependencies = TRUE, type = ifelse(tolower(Sys.info()["sysname"]) == "linux", "source", "binary"))
+          remotes::install_deps(dependencies = TRUE, type = "binary")
           remotes::install_dev("pkgdown")
         shell: Rscript {0}
 
