+++ vignettes/mcmc_rep.Rmd
@@ -29,7 +29,7 @@ library(jagstargets)
 
 It is sometimes desirable to run one or more Bayesian models repeatedly across multiple simulated datasets. Examples:
 
-1. Validate the implementation of a Bayesian model using simulation-based calibration [SBC; @cook2006; @talts2020].
+1. Validate the implementation of a Bayesian model using simulation [@cook2006; @talts2020].
 2. Simulate a randomized controlled experiment to explore frequentist properties such as power and Type I error.
 
 This vignette focuses on (1). The goal is to simulate multiple datasets from the model below, analyze each dataset, and assess how often the estimated credible intervals for `beta` capture the true value of `beta` from the simulation.
