### 2021-03-09 067bd97
Use .join_data instead of copy_data
