+++ R/tar_jags.R
@@ -23,7 +23,10 @@
 #' @inheritParams targets::tar_target
 #' @param name Symbol, base name for the collection of targets.
 #'   Serves as a prefix for target names.
-#' @param data Code to generate the `data` for the JAGS model.
+#' @param data Code to generate the `data` list for the JAGS model.
+#'   Optionally include a `.join_data` element to join parts of the data
+#'   to correspondingly named parameters in the summary output.
+#'   See the vignettes for details.
 #' @param jags_files Character vector of JAGS model files. If you
 #'   supply multiple files, each model will run on the one shared dataset
 #'   generated by the code in `data`. If you supply an unnamed vector,
@@ -129,25 +132,27 @@ tar_jags <- function(
     tidy_eval = tidy_eval
   )
   command_draws <- substitute(
-    jagstargets::tar_jags_df(fit, output = "draws"),
-    env = list(fit = sym_mcmc)
+    jagstargets::tar_jags_df(fit = fit, data = data, output = "draws"),
+    env = list(fit = sym_mcmc, data = sym_data)
   )
   command_summary <- substitute(
     jagstargets::tar_jags_df(
       fit,
+      data = data,
       output = "summary",
       summaries = quote(summaries),
       summary_args = quote(summary_args)
     ),
     env = list(
       fit = sym_mcmc,
+      data = sym_data,
       summaries = substitute(summaries),
       summary_args = substitute(summary_args)
     )
   )
   command_dic <- substitute(
-    jagstargets::tar_jags_df(fit, output = "dic"),
-    env = list(fit = sym_mcmc)
+    jagstargets::tar_jags_df(fit = fit, data = data, output = "dic"),
+    env = list(fit = sym_mcmc, data = sym_data)
   )
   args_mcmc <- list(
     call_ns("jagstargets", "tar_jags_run"),
@@ -319,10 +324,12 @@ tar_jags_run <- function(
     on.exit(sink(file = NULL, type = "output"))
   }
   lapply(jags.module, rjags::load.module, quiet = TRUE)
+  jags_data <- data
+  jags_data$.join_data <- NULL
   trn(
     n.cluster > 1L,
     R2jags::jags.parallel(
-      data = data,
+      data = jags_data,
       inits = inits,
       parameters.to.save = parameters.to.save,
       model.file = file,
@@ -338,7 +345,7 @@ tar_jags_run <- function(
       envir = envir
     ),
     R2jags::jags(
-      data = data,
+      data = jags_data,
       inits = inits,
       parameters.to.save = parameters.to.save,
       model.file = file,
