+++ tests/testthat/test-tar_jags_rep_summary.R
@@ -10,9 +10,7 @@ targets::tar_test("tar_jags_rep_summary()", {
       tar_jags_rep_summary(
         model,
         jags_files = c(x = "a.jags", y = "b.jags"),
-        data = tar_jags_example_data(),
-        data_copy = "true_beta",
-        data_omit = "true_beta",
+       data = tar_jags_example_data(),
         parameters.to.save = "beta",
         variables = "beta",
         log = R.utils::nullfile(),
@@ -67,12 +65,11 @@ targets::tar_test("tar_jags_rep_summary()", {
   expect_equal(length(out$y), 10L)
   expect_true(is.numeric(out$x))
   expect_true(is.numeric(out$y))
-  expect_true(is.numeric(out$true_beta))
   out1 <- targets::tar_read(model_x)
   out2 <- targets::tar_read(model_y)
   out <- targets::tar_read(model)
   expect_true("beta" %in% out$variable)
-  expect_true(all(c("true_beta", "mean", "q5") %in% colnames(out)))
+  expect_true(all(c("mean", "q5") %in% colnames(out)))
   expect_equal(sort(unique(out$.file)), sort(unique(c("a.jags", "b.jags"))))
   expect_equal(sort(unique(out$.name)), sort(unique(c("x", "y"))))
   expect_equal(dplyr::bind_rows(out1, out2), out)
@@ -98,8 +95,6 @@ targets::tar_test("tar_jags_rep_summary()", {
         model,
         jags_files = c(x = "a.jags", y = "b.jags"),
         data = c(tar_jags_example_data()),
-        data_copy = "true_beta",
-        data_omit = "true_beta",
         parameters.to.save = "beta",
         variables = "beta",
         log = R.utils::nullfile(),
@@ -137,8 +132,6 @@ targets::tar_test("tar_jags_rep_summary() with custom summaries", {
         model,
         jags_files = c(x = "a.jags", y = "b.jags"),
         data = tar_jags_example_data(),
-        data_copy = "true_beta",
-        data_omit = "true_beta",
         parameters.to.save = "beta",
         log = R.utils::nullfile(),
         refresh = 0,
@@ -160,6 +153,49 @@ targets::tar_test("tar_jags_rep_summary() with custom summaries", {
   capture.output(targets::tar_make(callr_function = NULL))
   out <- targets::tar_read(model)
   expect_true("q30" %in% colnames(out))
-  expect_true("true_beta" %in% colnames(out))
   expect_true(all(out$custom2 == 34))
+})
+
+targets::tar_test("join to summaries", {
+  skip_on_cran()
+  skip_if_not_installed("dplyr")
+  path <- system.file(
+    "join_data.jags",
+    package = "jagstargets",
+    mustWork = TRUE
+  )
+  file.copy(path, "a.jags")
+  targets::tar_script({
+    sim_data <- function (n = 10L) {
+      alpha <- stats::rnorm(n = 1, mean = 0, sd = 1)
+      beta <- stats::rnorm(n = n, mean = 0, sd = 1)
+      x <- seq(from = -1, to = 1, length.out = n)
+      y <- stats::rnorm(n, x * beta, 1)
+      .join_data = list(alpha = alpha, beta = beta)
+      out <- list(n = n, x = x, y = y, .join_data = .join_data)
+      out
+    }
+    list(
+      tar_jags_rep_summary(
+        model,
+        jags_files = "a.jags",
+        data = sim_data(),
+        parameters.to.save = c("alpha", "beta"),
+        log = R.utils::nullfile(),
+        refresh = 0,
+        n.iter = 2e3,
+        n.burnin = 1e3,
+        n.thin = 1,
+        n.chains = 4,
[…]
