+++ NEWS.md
@@ -1,0 +1,7 @@
+# jagstargets 0.0.0.9001
+
+* Use `.join_data` list element instead of arguments `copy_data` and `omit_data`.
+
+# jagstargets 0.0.0.9000
+
+* Added a `NEWS.md` file to track changes to the package.
