+++ R/tar_jags.R
@@ -52,7 +52,7 @@
 #'   tar_jags(
 #'     your_model,
 #'     jags_files = tmp,
-#'     data = tar_jags_example_data(true_params = FALSE),
+#'     data = tar_jags_example_data(),
 #'     parameters.to.save = "beta",
 #'     log = R.utils::nullfile()
 #'   )
