+++ R/tar_jags_rep_dic.R
@@ -34,7 +34,7 @@
 #'   tar_jags_rep_dic(
 #'     your_model,
 #'     jags_files = tmp,
-#'     data = tar_jags_example_data(true_params = FALSE),
+#'     data = tar_jags_example_data(),
 #'     parameters.to.save = "beta",
 #'     batches = 2,
 #'     reps = 2,
