+++ R/tar_jags_rep_summary.R
@@ -35,7 +35,7 @@
 #'   tar_jags_rep_summary(
 #'     your_model,
 #'     jags_files = tmp,
-#'     data = tar_jags_example_data(true_params = FALSE),
+#'     data = tar_jags_example_data(),
 #'     parameters.to.save = "beta",
 #'     batches = 2,
 #'     reps = 2,
