+++ man/tar_jags_rep_dic.Rd
@@ -243,7 +243,7 @@ list(
   tar_jags_rep_dic(
     your_model,
     jags_files = tmp,
-    data = tar_jags_example_data(true_params = FALSE),
+    data = tar_jags_example_data(),
     parameters.to.save = "beta",
     batches = 2,
     reps = 2,
