+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: dd94073575aaee2fd480892d24dfe0be85f81a03
 articles:
   mcmc: mcmc.html
   mcmc_rep: mcmc_rep.html
-last_built: 2021-03-09T13:44Z
+last_built: 2021-03-09T15:52Z
 
