### 2021-03-11 f29c95f
Edit mcmc_rep.Rmd vignette
