+++ vignettes/mcmc_rep.Rmd
@@ -29,10 +29,10 @@ library(jagstargets)
 
 It is sometimes desirable to run one or more Bayesian models repeatedly across multiple simulated datasets. Examples:
 
-1. Validate the implementation of a Bayesian model using simulation [@cook2006; @talts2020].
+1. Validate the implementation of a Bayesian model using simulation.
 2. Simulate a randomized controlled experiment to explore frequentist properties such as power and Type I error.
 
-This vignette focuses on (1). The goal is to simulate multiple datasets from the model below, analyze each dataset, and assess how often the estimated credible intervals of true values of parameters from the simulation.
+This vignette focuses on (1). The goal of this particular example to simulate multiple datasets from the model below, analyze each dataset, and assess how often the estimated posterior intervals cover the true parameters from the prior predictive simulations. The quantile method by @cook2006 generalizes this concept, and simulation-based calibration [@talts2020] generalizes further. The interval-based technique featured in this vignette is not as robust as SBC, but it may be more expedient for large models because it does not require visual inspection of multiple histograms.
 
 ```{r}
 lines <- "model {
