### 2021-03-18 23e632e
Change log to stdout and stderr
