+++ tests/testthat/test-tar_jags_rep_summary.R
@@ -13,7 +13,8 @@ targets::tar_test("tar_jags_rep_summary()", {
        data = tar_jags_example_data(),
         parameters.to.save = "beta",
         variables = "beta",
-        log = R.utils::nullfile(),
+        stdout = R.utils::nullfile(),
+        stderr = R.utils::nullfile(),
         refresh = 0,
         n.iter = 2e3,
         n.burnin = 1e3,
@@ -97,7 +98,8 @@ targets::tar_test("tar_jags_rep_summary()", {
         data = c(tar_jags_example_data()),
         parameters.to.save = "beta",
         variables = "beta",
-        log = R.utils::nullfile(),
+        stdout = R.utils::nullfile(),
+        stderr = R.utils::nullfile(),
         refresh = 0,
         n.iter = 2e3,
         n.burnin = 1e3,
@@ -133,7 +135,8 @@ targets::tar_test("tar_jags_rep_summary() with custom summaries", {
         jags_files = c(x = "a.jags", y = "b.jags"),
         data = tar_jags_example_data(),
         parameters.to.save = "beta",
-        log = R.utils::nullfile(),
+        stdout = R.utils::nullfile(),
+        stderr = R.utils::nullfile(),
         refresh = 0,
         n.iter = 2e3,
         n.burnin = 1e3,
@@ -181,7 +184,8 @@ targets::tar_test("join to summaries", {
         jags_files = "a.jags",
         data = sim_data(),
         parameters.to.save = c("alpha", "beta"),
-        log = R.utils::nullfile(),
+        stdout = R.utils::nullfile(),
+        stderr = R.utils::nullfile(),
         refresh = 0,
         n.iter = 2e3,
         n.burnin = 1e3,
