+++ tests/testthat/test-tar_jags_rep_dic.R
@@ -12,7 +12,8 @@ targets::tar_test("tar_jags_rep_dic()", {
         jags_files = c(x = "a.jags", y = "b.jags"),
         data = tar_jags_example_data(),
         parameters.to.save = "beta",
-        log = R.utils::nullfile(),
+        stdout = R.utils::nullfile(),
+        stderr = R.utils::nullfile(),
         refresh = 0,
         n.iter = 2e3,
         n.burnin = 1e3,
@@ -94,7 +95,8 @@ targets::tar_test("tar_jags_rep_dic()", {
         jags_files = c(x = "a.jags", y = "b.jags"),
         data = c(tar_jags_example_data()),
         parameters.to.save = "beta",
-        log = R.utils::nullfile(),
+        stdout = R.utils::nullfile(),
+        stderr = R.utils::nullfile(),
         refresh = 0,
         n.iter = 2e3,
         n.burnin = 1e3,
