### 2021-03-24 2b08dfe
Switch meanings of logical infix ops to conform to precedent
