+++ R/tar_jags_df.R
@@ -45,7 +45,7 @@ tar_jags_df <- function(
 tar_jags_df_draws <- function(fit) {
   list <- fit$BUGSoutput$sims.list
   out <- as.matrix(fit$BUGSoutput$sims.matrix, nrow = nrow(list))
-  colnames(out) <- colnames(out) %|||% names(list)
+  colnames(out) <- colnames(out) %||% names(list)
   out <- tibble::as_tibble(out)
   chains <- as.integer(fit$BUGSoutput$n.chains)
   iterations <- as.integer(nrow(out) / chains)
