+++ R/utils_assert.R
@@ -1,30 +1,30 @@
 assert_chr <- function(x, msg = NULL) {
   if (!is.character(x)) {
-    throw_validate(msg %||% "x must be a character.")
+    throw_validate(msg %|||% "x must be a character.")
   }
 }
 
 assert_in <- function(x, choices, msg = NULL) {
   if (!all(x %in% choices)) {
-    throw_validate(msg %||% paste(deparse(x), "is in", deparse(choices)))
+    throw_validate(msg %|||% paste(deparse(x), "is in", deparse(choices)))
   }
 }
 
 assert_not_in <- function(x, choices, msg = NULL) {
   if (any(x %in% choices)) {
-    throw_validate(msg %||% paste(deparse(x), "is in", deparse(choices)))
+    throw_validate(msg %|||% paste(deparse(x), "is in", deparse(choices)))
   }
 }
 
 assert_nzchar <- function(x, msg = NULL) {
   if (any(!nzchar(x))) {
-    throw_validate(msg %||% "x has empty character strings")
+    throw_validate(msg %|||% "x has empty character strings")
   }
 }
 
 assert_package <- function(package, msg = NULL) {
   if (!requireNamespace(package, quietly = TRUE)) {
-    throw_validate(msg %||% paste("package ", package, " not installed"))
+    throw_validate(msg %|||% paste("package ", package, " not installed"))
   }
 }
 
@@ -32,7 +32,7 @@ assert_path <- function(path, msg = NULL) {
   missing <- !file.exists(path)
   if (any(missing)) {
     throw_validate(
-      msg %||% paste(
+      msg %|||% paste(
         "missing files: ",
         paste(path[missing], collapse = ", ")
       )
@@ -42,7 +42,7 @@ assert_path <- function(path, msg = NULL) {
 
 assert_scalar <- function(x, msg = NULL) {
   if (length(x) != 1) {
-    throw_validate(msg %||% "x must have length 1.")
+    throw_validate(msg %|||% "x must have length 1.")
   }
 }
 
@@ -55,6 +55,6 @@ assert_jags_file <- function(jags_file) {
 assert_unique <- function(x, msg = NULL) {
   if (anyDuplicated(x)) {
     dups <- paste(unique(x[duplicated(x)]), collapse = ", ")
-    throw_validate(paste(msg %||% "duplicated entries:", dups))
+    throw_validate(paste(msg %|||% "duplicated entries:", dups))
   }
 }
