+++ R/utils_language.R
@@ -31,7 +31,7 @@ produce_direct_deparse <- function() {
   deparse
 }
 
-tidy_eval <- function(expr, envir, tidy_eval) {
+tar_tidy_eval <- function(expr, envir, tidy_eval) {
   if (tidy_eval) {
     expr <- as.call(c(quote(rlang::expr), expr))
     expr <- rlang::quo_squash(eval(expr, envir = envir))
