### 2021-03-25 d38b9b9
Minor config/dep updates
