+++ R/utils_assert.R
@@ -10,6 +10,12 @@ assert_in <- function(x, choices, msg = NULL) {
   }
 }
 
+assert_nonempty <- function(x, msg = NULL) {
+  if (!length(x)) {
+    throw_validate(msg %|||% "x must not be empty")
+  }
+}
+
 assert_not_in <- function(x, choices, msg = NULL) {
   if (any(x %in% choices)) {
     throw_validate(msg %|||% paste(deparse(x), "is in", deparse(choices)))
@@ -22,13 +28,8 @@ assert_nzchar <- function(x, msg = NULL) {
   }
 }
 
-assert_package <- function(package, msg = NULL) {
-  if (!requireNamespace(package, quietly = TRUE)) {
-    throw_validate(msg %|||% paste("package ", package, " not installed"))
-  }
-}
-
 assert_path <- function(path, msg = NULL) {
+  assert_nonempty(path, "path must not be empty")
   missing <- !file.exists(path)
   if (any(missing)) {
     throw_validate(
@@ -50,6 +51,9 @@ assert_jags_file <- function(jags_file) {
   assert_scalar(jags_file, "only one jags model file allowed at a time.")
   assert_chr(jags_file, "jags_file must be a character.")
   assert_path(jags_file, "jags_file is not an existing file.")
+  if (dir.exists(jags_file)) {
+    throw_validate("jags_file must not be a directory.")
+  }
 }
 
 assert_unique <- function(x, msg = NULL) {
