+++ R/tar_jags_rep.R
@@ -72,14 +72,14 @@ tar_jags_rep <- function(
   name_lines <- paste0(name, "_lines")
   name_batch <- paste0(name, "_batch")
   name_data <- paste0(name, "_data")
-  sym_jags <- rlang::syms(name_jags)
-  sym_file <- rlang::sym(name_file)
-  sym_lines <- rlang::sym(name_lines)
-  sym_batch <- rlang::sym(name_batch)
-  sym_data <- rlang::sym(name_data)
+  sym_jags <- as_symbols(name_jags)
+  sym_file <- as.symbol(name_file)
+  sym_lines <- as.symbol(name_lines)
+  sym_batch <- as.symbol(name_batch)
+  sym_data <- as.symbol(name_data)
   command_lines <- call_function(
     "readLines",
-    args = list(con = rlang::sym(name_file))
+    args = list(con = as.symbol(name_file))
   )
   command_batch <- substitute(seq_len(x), env = list(x = batches))
   command_rep <- tar_tidy_eval(data, envir = envir, tidy_eval = tidy_eval)
