+++ R/tar_jags_package.R
@@ -18,7 +18,6 @@
 #' @importFrom qs qread
 #' @importFrom R2jags jags jags.parallel
 #' @importFrom rjags load.module
-#' @importFrom rlang as_function sym
 #' @importFrom stats rnorm runif
 #' @importFrom targets tar_cue tar_dir tar_load tar_option_get tar_path
 #'   tar_read tar_script tar_target tar_target_raw tar_test
