+++ NAMESPACE
@@ -20,8 +20,6 @@ importFrom(purrr,map)
 importFrom(purrr,map_dbl)
 importFrom(qs,qread)
 importFrom(rjags,load.module)
-importFrom(rlang,as_function)
-importFrom(rlang,sym)
 importFrom(stats,rnorm)
 importFrom(stats,runif)
 importFrom(tarchetypes,tar_combine)
