+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: 9a5a8f5cd29bb06616c1452d1b9b5cab7cebf7c2
 articles:
   mcmc: mcmc.html
   mcmc_rep: mcmc_rep.html
-last_built: 2021-03-25T18:00Z
+last_built: 2021-03-25T18:05Z
 
