+++ NAMESPACE
@@ -20,6 +20,8 @@ importFrom(purrr,map)
 importFrom(purrr,map_dbl)
 importFrom(qs,qread)
 importFrom(rjags,load.module)
+importFrom(rlang,expr)
+importFrom(rlang,quo_squash)
 importFrom(stats,rnorm)
 importFrom(stats,runif)
 importFrom(tarchetypes,tar_combine)
