+++ README.Rmd
@@ -46,7 +46,7 @@ And if you have not done so already, install JAGS from <https://mcmc-jags.source
 
 ## Usage
 
-First, write a [`_targets.R` file](https://books.ropensci.org/targets/walkthrough.html) that loads your packages, defines a function to generate JAGS data, and lists a pipeline of targets. The target list can call target factories like [`tar_jags()`](https://wlandau.github.io/jagstargets/reference/tar_jags.html) as well as ordinary targets with [`tar_target()`](https://docs.ropensci.org/targets/reference/tar_target.html).
+First, write a [`_targets.R` file](https://books.ropensci.org/targets/walkthrough.html) that loads your packages, defines a function to generate JAGS data, and lists a pipeline of targets. The target list can call target factories like [`tar_jags()`](https://wlandau.github.io/jagstargets/reference/tar_jags.html) as well as ordinary targets with [`tar_target()`](https://docs.ropensci.org/targets/reference/tar_target.html). The following minimal example is simple enough to contain entirely within the `_targets.R` file, but for larger projects, you may wish to store functions in separate files as in the [`targets-keras`](https://github.com/wlandau/targets-keras) example.
 
 ```{r, eval = FALSE}
 # _targets.R
@@ -78,6 +78,6 @@ Development is a community effort, and we welcome discussion and contribution. B
 
 ## Citation
 
-```{r}
+```{r, warning = FALSE}
 citation("jagstargets")
 ```
