+++ R/tar_jags_example_data.R
@@ -10,6 +10,19 @@
 #'   element so the true value of `beta` from the simulation
 #'   is automatically appended to the `beta` rows of the
 #'   summary output.
+#' @format A list with the following elements:
+#'   * `n`: integer, number of data points.
+#'   * `x`: numeric, covariate vector.
+#'   * `y`: numeric, response variable.
+#'   * `true_beta`: numeric of length 1, value of the regression
+#'     coefficient `beta` used in simulation.
+#'   * `.join_data`: a list of simulated values to be appended
+#'     to as a `.join_data` column in the output of
+#'     targets generated by functions such as
+#'     [tar_jags_rep_summary()]. Contains the
+#'     regression coefficient `beta` (numeric of length 1)
+#'     and prior predictive data `y` (numeric vector).
+#' @param n Integer of length 1, number of data points.
 #' @param n Integer of length 1, number of data points.
 #' @examples
 #' tar_jags_example_data()
