+++ man/tar_jags_example_data.Rd
@@ -3,6 +3,22 @@
 \name{tar_jags_example_data}
 \alias{tar_jags_example_data}
 \title{Example data for \code{\link[=tar_jags_example_file]{tar_jags_example_file()}}}
+\format{
+A list with the following elements:
+\itemize{
+\item \code{n}: integer, number of data points.
+\item \code{x}: numeric, covariate vector.
+\item \code{y}: numeric, response variable.
+\item \code{true_beta}: numeric of length 1, value of the regression
+coefficient \code{beta} used in simulation.
+\item \code{.join_data}: a list of simulated values to be appended
+to as a \code{.join_data} column in the output of
+targets generated by functions such as
+\code{\link[=tar_jags_rep_summary]{tar_jags_rep_summary()}}. Contains the
+regression coefficient \code{beta} (numeric of length 1)
+and prior predictive data \code{y} (numeric vector).
+}
+}
 \usage{
 tar_jags_example_data(n = 10L)
 }
