+++ README.Rmd
@@ -31,7 +31,7 @@ The `jagstargets` R package is an extension to [`targets`](https://docs.ropensci
 ## Prerequisites
 
 1. The [prerequisites of the `targets` R package](https://docs.ropensci.org/targets/#prerequisites).
-1. Basic familiarity with [`targets`](https://docs.ropensci.org/targets/): watch minutes 6 through 40 of [this video](https://youtu.be/Gqn7Xn4d5NI), then read [this chapter](https://books.ropensci.org/targets/walkthrough.html) of the [user manual](https://books.ropensci.org/targets/).
+1. Basic familiarity with [`targets`](https://docs.ropensci.org/targets/): watch minutes 7 through 40 of [this video](https://youtu.be/Gqn7Xn4d5NI?t=439), then read [this chapter](https://books.ropensci.org/targets/walkthrough.html) of the [user manual](https://books.ropensci.org/targets/).
 1. Familiarity with Bayesian Statistics and [JAGS](https://mcmc-jags.sourceforge.io/). Prior knowledge of [`rjags`](https://cran.r-project.org/package=rjags) or [`R2jags`](https://cran.r-project.org/package=R2jags) helps.
 
 ## How to get started
