+++ DESCRIPTION
@@ -32,27 +32,27 @@ Depends:
   R (>= 3.5.0)
 Imports:
   coda (>= 0.19.4),
-  digest (>= 0.6.27),
-  fst (>= 0.9.4),
-  posterior (>= 0.1.3),
+  digest (>= 0.6.25),
+  fst (>= 0.9.2),
+  posterior (>= 0.1.2),
   purrr (>= 0.3.4),
-  qs (>= 0.14.1),
+  qs (>= 0.23.2),
   R2jags (>= 0.6.1),
   rlang (>= 0.4.10),
   rjags (>= 4.10),
   stats,
   targets (>= 0.0.1),
   tarchetypes (>= 0.0.1),
-  tibble (>= 3.0.4),
+  tibble (>= 3.0.1),
   tools,
   utils,
-  withr (>= 2.3.0),
+  withr (>= 2.1.2),
 Suggests:
   dplyr (>= 1.0.2),
   fs (>= 1.5.0),
-  knitr (>= 1.28),
+  knitr (>= 1.30),
   R.utils (>= 2.10.1),
-  rmarkdown (>= 2.1),
+  rmarkdown (>= 2.3),
   testthat (>= 3.0.0),
   tidyr (>= 1.1.2),
   visNetwork (>= 2.0.9)
