+++ R/tar_jags.R
@@ -1,4 +1,4 @@
-#' @title One MCMC per model with multiple outputs.
+#' @title One MCMC per model with multiple outputs
 #' @export
 #' @description Targets to run a JAGS model once with MCMC
 #'   and save multiple outputs.
