+++ NAMESPACE
@@ -18,6 +18,7 @@ importFrom(fst,read_fst)
 importFrom(posterior,summarize_draws)
 importFrom(purrr,map)
 importFrom(purrr,map_dbl)
+importFrom(purrr,map_dfr)
 importFrom(qs,qread)
 importFrom(rjags,load.module)
 importFrom(rlang,expr)
