+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: 5b5baea274400b242e0a2194fe9d392f753d3508
 articles:
   mcmc: mcmc.html
   mcmc_rep: mcmc_rep.html
-last_built: 2021-04-04T13:32Z
+last_built: 2021-04-04T13:39Z
 
