### 2021-04-04 abbd513
Check for JAGS files to start
