+++ R/tar_jags_rep_dic.R
@@ -38,7 +38,7 @@
 #' library(jagstargets)
 #' # Do not use a temp file for a real project
 #' # or else your targets will always rerun.
-#' tmp <- tempfile()
+#' tmp <- tempfile(pattern = "", fileext = ".jags")
 #' tar_jags_example_file(tmp)
 #' list(
 #'   tar_jags_rep_dic(
@@ -94,6 +94,7 @@ tar_jags_rep_dic <- function(
   retrieval = targets::tar_option_get("retrieval"),
   cue = targets::tar_option_get("cue")
 ) {
+  lapply(jags_files, assert_jags_file)
   tar_jags_rep(
     name = deparse_language(substitute(name)),
     jags_files = jags_files,
