+++ R/tar_jags_example_file.R
@@ -5,10 +5,12 @@
 #' @return `NULL` (invisibly).
 #' @param path Character of length 1, file path to write the model file.
 #' @examples
-#' path <- tempfile()
+#' path <- tempfile(pattern = "", fileext = ".jags")
 #' tar_jags_example_file(path = path)
 #' writeLines(readLines(path))
-tar_jags_example_file <- function(path = tempfile()) {
+tar_jags_example_file <- function(
+  path = tempfile(pattern = "", fileext = ".jags")
+) {
   src <- system.file("example.jags", package = "jagstargets", mustWork = TRUE)
   file.copy(src, path, overwrite = TRUE)
   invisible()
