### 2021-04-05 665dd49
trn() => if_any()
