+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -3,7 +3,6 @@
 * [ ] I understand and agree to this repository's [code of conduct](https://github.com/wlandau/jagstargets/blob/main/CODE_OF_CONDUCT.md).
 * [ ] I understand and agree to this repository's [contributing guidelines](https://github.com/wlandau/jagstargets/blob/main/CONTRIBUTING.md).
 * [ ] I have already submitted an issue to the [issue tracker](http://github.com/wlandau/jagstargets/issues) to discuss my idea with the maintainer.
-* [ ] This pull request is not a [draft](https://github.blog/2019-02-14-introducing-draft-pull-requests).
 
 # Related GitHub issues and pull requests
 
