### 2021-04-14 #425 @noamross
Hi @wlandau!  Running editor checks, I am able to fully test the package but have the following issue checking the package using `devtools::check()`:

```
✓  checking for file ‘/home/noamross/reviews/jagstargets/DESCRIPTION’ ...
─  preparing ‘jagstargets’:
✓  checking DESCRIPTION meta-information ...
E  creating vignettes (30.7s)
   --- re-building ‘mcmc.Rmd’ using rmarkdown
   Quitting from lines 284-285 (mcmc.Rmd)
   Error: processing vignette 'mcmc.Rmd' failed with diagnostics:
   callr subprocess failed: jags_file is not an existing file.
   Visit https://books.ropensci.org/targets/debugging.html for debugging advice.
   --- failed re-building ‘mcmc.Rmd’

   --- re-building ‘mcmc_rep.Rmd’ using rmarkdown
   Quitting from lines 354-355 (mcmc_rep.Rmd)
   Error: processing vignette 'mcmc_rep.Rmd' failed with diagnostics:
   callr subprocess failed: jags_file is not an existing file.
   Visit https://books.ropensci.org/targets/debugging.html for debugging advice.
   --- failed re-building ‘mcmc_rep.Rmd’

   SUMMARY: processing the following files failed:
     ‘mcmc.Rmd’ ‘mcmc_rep.Rmd’
```

This appears to occur the `visNetwork()` call in the vignettes (commenting them out allows the vignettes to build). Otherwise all tests pass and I am running JAGS 4.3.0.
