+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: 3dd1a1ad7740977491241c89eee15d9b58e75866
 articles:
   mcmc: mcmc.html
   mcmc_rep: mcmc_rep.html
-last_built: 2021-04-14T18:09Z
+last_built: 2021-04-14T18:19Z
 
