### 2021-04-15 #425 @noamross
### Editor checks:

- [x] **Automated tests:** Package has a testing suite and is tested via a CI service.

---

Thank you will. All tests are running fine now, and your `goodpractice::gp()` diagnostics are all clear! I am seeking reviewers.
