### 2021-04-29 #425 @dill
hi @wlandau, I'm just getting started here and in installation I see that [currently `targets` has been archived on CRAN](https://cloud.r-project.org/web/packages/targets/index.html) (which I assume you're already aware of) so I have installed the version from [`ropensci/targets`](https://github.com/ropensci/targets). Is that the most appropriate version to work from?
