### 2021-04-29 #425 @wlandau
Thanks for reviewing, @dill. Development `targets` should be fine while we wait for CRAN to review `targets` version 0.4.2, which I submitted last weekend.
