### 2021-04-29 wlandau/jagstargets#14 @dill
"Duplicate" chunks in MCMC pipelines vignette
## Prework

* [X] Read and agree to the [code of conduct](https://ropensci.org/code-of-conduct/) and [contributing guidelines](https://github.com/ropensci/targets/blob/main/CONTRIBUTING.md).
* [X] If there is [already a relevant issue](https://github.com/ropensci/targets/issues), whether open or closed, comment on the existing thread instead of posting a new issue.

## Description

This is part of ropensci/software-review#425, please don't feel you need to respond to this until I have submitted my full review.

In vignette `mcmc.Rmd` there are a few places ([here](https://github.com/wlandau/jagstargets/blob/fc3154d490544480c42bb6f41b1ce9fb96262acb/vignettes/mcmc.Rmd#L47), [here](https://github.com/wlandau/jagstargets/blob/fc3154d490544480c42bb6f41b1ce9fb96262acb/vignettes/mcmc.Rmd#L150) and [here](https://github.com/wlandau/jagstargets/blob/fc3154d490544480c42bb6f41b1ce9fb96262acb/vignettes/mcmc.Rmd#L226)) where there are basically two copies of the code, one that is actually `eval`ed and one that's there to render in the output doco. That's totally fine, but the un`echo`'d chunk might be improved with some comments explaining why there are two versions of this chunk, for those following along in their own `.Rmd` or using your vignette as a template (and when one might need to do this in our own docos).
