+++ vignettes/mcmc.Rmd
@@ -66,7 +66,7 @@ tar_script({
 library(targets)
 library(jagstargets)
 
-generate_data <- function() {
+generate_data <- function(n = 10) {
   true_beta <- stats::rnorm(n = 1, mean = 0, sd = 1)
   x <- seq(from = -1, to = 1, length.out = n)
   y <- stats::rnorm(n, x * true_beta, 1)
@@ -177,7 +177,7 @@ tar_script({
 library(targets)
 library(jagstargets)
 
-generate_data <- function() {
+generate_data <- function(n = 10) {
   true_beta <- stats::rnorm(n = 1, mean = 0, sd = 1)
   x <- seq(from = -1, to = 1, length.out = n)
   y <- stats::rnorm(n, x * true_beta, 1)
@@ -253,7 +253,7 @@ file.copy("x.jags", "y.jags")
 library(targets)
 library(jagstargets)
 
-generate_data <- function() {
+generate_data <- function(n = 10) {
   true_beta <- stats::rnorm(n = 1, mean = 0, sd = 1)
   x <- seq(from = -1, to = 1, length.out = n)
   y <- stats::rnorm(n, x * true_beta, 1)
