+++ vignettes/mcmc_rep.Rmd
@@ -48,6 +48,7 @@ writeLines(lines, "model.jags")
 Next, we define a pipeline to simulate multiple datasets and fit each dataset with the model. In our data-generating function, we put the true parameter values of each simulation in a special `.join_data` list. `jagstargets` will automatically join the elements of `.join_data` to the correspondingly named variables in the summary output. This will make it super easy to check how often our posterior intervals capture the truth. As for scale, we declare 10 replications: 2 batches with 5 iterations per batch. (In practical situations, the total number of replications should be hundreds of times more.) 
 
 ```{r, echo = FALSE}
+# Writes the _targets.R file shown in the next code chunk.
 library(targets)
 tar_script({
   library(jagstargets)
@@ -148,6 +149,7 @@ model %>%
 For maximum reproducibility, we should express the coverage assessment as a custom function and a target in the pipeline.
 
 ```{r, echo = FALSE}
+# Writes the _targets.R file shown in the next code chunk.
 library(targets)
 tar_script({
   library(jagstargets)
@@ -262,6 +264,7 @@ tar_read(coverage)
 
 
 ```{r, echo = FALSE}
+# Writes the _targets.R file shown in the next code chunk.
 library(targets)
 tar_script({
   library(jagstargets)
