### 2021-04-29 wlandau/jagstargets#14 @wlandau
Sure, I added some comments in https://github.com/wlandau/jagstargets/commit/18eeb2db40a49e5031868dac8f145095bd6135a4. `targets` requires a `_targets.R` file to run the analysis, but `tar_script()` in a visible code chunk is awkward to read. Having 2 code chunks is the only way I have been able to find that lets us have both automation and readability.
