+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: dff8062d297c001dba0dfaf567bfffdc8ad89f5c
 articles:
   mcmc: mcmc.html
   mcmc_rep: mcmc_rep.html
-last_built: 2021-04-29T19:30Z
+last_built: 2021-04-29T19:36Z
 
