+++ DESCRIPTION
@@ -57,6 +57,8 @@ Suggests:
   tidyr (>= 1.1.2),
   visNetwork (>= 2.0.9)
 Remotes:
+  ropensci/targets,
+  ropensci/tarchetypes,
   stan-dev/posterior
 SystemRequirements: JAGS 4.x.y (http://mcmc-jags.sourceforge.net)
 Encoding: UTF-8
