### 2021-05-06 wlandau/jagstargets#17 @dill
`generate_data` in `mcmc_rep.Rmd` doesn't use `alpha`?
## Prework

* [X] Read and agree to the [code of conduct](https://github.com/wlandau/jagstargets/blob/main/CODE_OF_CONDUCT.md) and [contributing guidelines](https://github.com/wlandau/jagstargets/blob/main/CONTRIBUTING.md).
* [X] Confirm that your issue is most likely a genuine bug in `jagstargets` and not a known limitation or usage error.
* [X] If there is [already a relevant issue](https://github.com/wlandau/jagstargets/issues), whether open or closed, comment on the existing thread instead of posting a new issue.
* [X] Post a [minimal reproducible example](https://www.tidyverse.org/help/) like [this one](https://github.com/wlandau/targets/issues/256#issuecomment-754229683) so the maintainer can troubleshoot the problems you identify. A reproducible example is:
    * [X] **Runnable**: post enough R code and data so any onlooker can create the error on their own computer.
    * [X] **Minimal**: reduce runtime wherever possible and remove complicated details that are irrelevant to the issue at hand.
    * [X] **Readable**: format your code according to the [tidyverse style guide](https://style.tidyverse.org/).

## Description

This is part of ropensci/software-review#425.

Running through the vignette [Scaling MCMC pipelines](https://wlandau.github.io/jagstargets/articles/mcmc_rep.html), the data is generated by the following function:

```r
generate_data <- function (n = 10L) {
  alpha <- stats::rnorm(n = 1, mean = 0, sd = 1)
  beta <- stats::rnorm(n = n, mean = 0, sd = 1)
  x <- seq(from = -1, to = 1, length.out = n)
  y <- stats::rnorm(n, x * beta, 1)
  # Elements of .join_data get joined on to the .join_data column
  # in the summary output next to the model parameters
  # with the same names.
  .join_data <- list(alpha = alpha, beta = beta)
  list(n = n, x = x, y = y, .join_data = .join_data)
}
```

Although `alpha` is generated, it is not actually used in the model. Should the line assigning `y` be:

```r
y <- stats::rnorm(n, alpha + x * beta, 1)
```

?

[link to line in this repo](https://github.com/wlandau/jagstargets/blob/e53192bf0f6bd772e2cc1d9f1f14150c73aa666f/vignettes/mcmc_rep.Rmd#L61)

## Diagnostic information

Issue exists in last commit 18eeb2db40a49e5031868dac8f145095bd6135a4
