### 2021-05-06 #425 @dill
- [X] **Examples** (that run successfully locally) for all exported functions
  - [X] `devtools::run_examples()`
  - [X] `devtools::run_examples(run_donttest=TRUE)`
  - [X] `devtools::run_examples(run_donttest=TRUE, run_dontrun=TRUE)`

I was using a 2020 MacBook Air (Intel) running macOS 10.15.7 (Catalina). Here is the output of `sessionInfo()`:

```
> sessionInfo()
Platform: x86_64-apple-darwin17.0 (64-bit)
Running under: macOS Catalina 10.15.7

BLAS:   /Library/Frameworks/R.framework/Versions/4.0/Resources/lib/libRblas.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/4.0/Resources/lib/libRlapack.dylib

[1] en_GB.UTF-8/en_GB.UTF-8/en_GB.UTF-8/C/en_GB.UTF-8/en_GB.UTF-8

[1] compiler_4.0.4
```

I tried:

```r
remotes::install_github("wlandau/jagstargets", dependencies=TRUE)
```

but (2021/04/29) `targets` had been removed from CRAN, so this failed with error:

```
ERROR: dependencies 'targets', 'tarchetypes' are not available for package 'jagstargets'
```

So instead I installed with

```r
remotes::install_github("wlandau/targets", dependencies=TRUE)
remotes::install_github("ropensci/tarchetypes", dependencies=TRUE)
remotes::install_github("wlandau/jagstargets", dependencies=TRUE)
```

This worked fine.

2021-05-06 I tried using `install_github()` as above again and had no problems.

  - [X] Unit tests cover essential functions of the package and a reasonable range of inputs and conditions.
    - `goodpractice::gp` reports: 17% of code lines are covered by test cases (this excludes running vignettes).
    - `covr::package_coverage(type="all")` reports: 99.57% coverage :+1:
  - [X] All tests pass on the local machine.

Estimated hours spent reviewing:  (this is high as it includes reading the rOpenSci review documentation)

---

For some background on where I'm coming from with this review: I am a statistician, but I work primarily with ecologists (who tend to be pretty keen on JAGS and JAGS-like programming), so I have some experience with these kinds of workflows. I don't have experience with `targets`, so that is new to me. I'm not an RStudio user, so this is based on using R from the command line or within [Nvim-R](https://github.com/jalvesaq/Nvim-R), I doubt this makes any difference but just an FYI.

I realise this is probably a daunting number of comments. In general `jagstargets` is an awesome piece of work! Most of my comments are aimed at making the package more accessible for people like me who want to get started using `jagstargets`. This was a fun package to review, and I learnt a lot. I hope the comments below are useful for you!

#### `README.md`

The first two paragraphs give a summary of the package but they feel like they are written for people who already know what `targets` does (e.g., "factories", "targets", "pipelines" are all jargon to begin with; though I appreciated the links to the blog post, I can see this turning people off right away). Perhaps there could be some adaptation so that someone who is a JAGS user can immediately understand how the package will save them time and headaches. Maybe even a "for JAGS users" and a "for people familiar with `targets`" section might help? (FWIW, I found the text in ?`jagstargets-package` to be a bit more transparent.)

- As stated above, all tests ran on my machine post installation. They took about 8s.
- Tests didn't use `context()` to provide extra information on what exactly was being tested each time. Filenames were semi-informative, but might be better to have more info there for contributors. This might be more useful if tests are expanded further.
- Not much in the way of comments in the test files. Some are self-explanatory but the large tests are harder to follow.

#### MCMC pipelines vignette

- Could this be renamed to "Getting started with..." or "An introduction to..." to make it clear this is where to start?
- It might be nice to have a bit of table setting at the start of this vignette to talk about the kinds of outputs we want, even showing how an existing JAGS workflow works and how that maps into `jagstargets` (and then you can say how similar this is to the DAG generated by `tar_visnetwork`? In order to get people on board I think it's necessary to show clear mapping between their current workflow and "doing something fancy" and a bit of intro here might help that happen.
- I found the list of targets a little confusing initially ([around here](https://github.com/wlandau/jagstargets/blob/fc3154d490544480c42bb6f41b1ce9fb96262acb/vignettes/mcmc.Rmd#L89)). I wasn't sure where I was supposed to be looking for those objects and what was going on. Perhaps running `tar_manifest()` first, showing the effect of the `tar_jags` call, then explaining what these entries mean might make that easier to understand? Does the `tar_jags` call have to be wrapped in `list()` in that chunk (if so maybe explain why).
- The last part of the vignette starts talking about a second model `y.jags` though we don't define that. Is it worth make just a quick note here that `tar_visnetwork` doesn't check that the model files exist?

#### Scaling MCMC pipelines vignette

- Might be worth having something up top saying to start with the other vignette first?
- There could be a little more chat at the start of this vignette explaining that you want to find out about how well `model.jags` estimates `alpha` and `beta[]`.
- This is a bit of an odd model (in my opinion at least!) in that you have fixed the covariate values and have `beta` be a random effect (you generate `n` random slopes), rather than a fixed `beta` with random data each time. I wonder if a classic linear model situation (where data vary and there is only one `beta`) would be a clearer situation for people to think about? It also has the advantage of producing much less output, as well as having the same DGP as the other vignette.
- With my stats hat on, I'd say write out the equations for the model/DGP, so it's super-clear what's happening here.
- There is some setup in the chunks here that isn't explained (e.g., `options(crayon.enabled = FALSE)`, `tar_option_set(memory = "transient", garbage_collection = TRUE)` [here](https://github.com/wlandau/jagstargets/blob/e53192bf0f6bd772e2cc1d9f1f14150c73aa666f/vignettes/mcmc_rep.Rmd#L55) that is worth explaining to novice users like myself. Perhaps this is only necessary for this vignette format, but it would be good to know that if so.
- I found the `_targets.R` definition hard to follow, particularly the arguments to `tar_jags_rep_summary()` were hard to relate to what is going on...
  - The terminology could be explained here at the same time: "branches", "replications", "batches" and "iterations" all appear almost at once. I think I've figured out what they mean by the end, but it would be nice to have that clear as those terms are and why we use them here. This could even be woven into the first 1-2 intro paragraphs?
- The names of the nodes in these graphs are hard-going (this was less obvious above when the model was simpler). `model_model` vs. `model` vs. `model_lines_model`: these make sense once you know what you're doing I'm sure but I found them difficult to keep track of (see comment in "Package code" below). Worth pointing people to the documentation in `?tar_jags` to explain the naming convention at this point.
- Is there any useful information that can be gleaned from the `.rep` column? It looks like a hash, can it tell us anything re: reproducibility or is it just for internal bookkeeping?

#### Trying this out for myself

I tried to construct a comparison between `mgcv::gam` and `mgcv::jagam` using `jagstargets`. You can find the resulting code [here](https://github.com/dill/jagam_jagstargets/blob/main/_targets.R). Mostly I found this fairly intuitive once I got my head around the above. `tar_visnetwork()` was essential to debugging. Feel no obligation to look at this but it might be useful to see what weird things people do when working from scratch and seeing where some bottlenecks are.

#### Package code

I can't offer too much comment here as the majority of the code is calls to `targets` which I am not really qualified to comment on in any depth. In general I found the code clear and understandable throughout, even though I don't know the ins and outs of `targets`.

- Per @wlandau's issue at MangoTheCat/goodpractice#129 (which turned up when I I google'd the error), I was unable to get all checks running (though I don't have the same issue with other packages?). This happened with both the CRAN and github versions of `goodpractice`, with the following message (obvs, this is not necessarily a problem with `jagstargets` but I thought I would report this here anyway):
```
1: In MYPREPS[[prep]](state, quiet = quiet) :
  Prep step for cyclomatic complexity failed.
2: In MYPREPS[[prep]](state, quiet = quiet) :
  Prep step for description failed.
3: In MYPREPS[[prep]](state, quiet = quiet) :
  Prep step for namespace failed.
```
- If `jagstargets-package` is a landing page, perhaps it could contain a link back to the `pkgdown` site, and perhaps a `@seealso` to the most useful starting manpage(s) (`tar_jags`?)
- `tar_jags_example_data.R` should explain the data generating process, as well as what the components of the returned object.
- Naming scheme for targets: are there more descriptive names for the `x_file_y` and `x_lines_y`? I don't have stellar suggestions on replacements but "file" and "lines" feel very generic and like they might get confused in a large workflow? The other labels seem very clear though.
- Following-on from this, it is possible to combine the `x_file_y` and `x_lines_y` targets? I see in [the `stantargets` vignette](https://docs.ropensci.org/stantargets/articles/mcmc.html) there doesn't seem to be a "compile" step, just the `x_file_y` then `x_mcmc_y` -- is there something that is different here that means we need both steps to be separate?

### Issues raised/pull requests

- [X] ["Duplicate" chunks in MCMC pipelines vignette #14](https://github.com/wlandau/jagstargets/issues/14)
- [X] [Missing definition of n in MCMC pipeline vignette #15 ](https://github.com/wlandau/jagstargets/issues/15)
- [X] [add @aliases for ?jagstargets-package to ?jagstargets #16](https://github.com/wlandau/jagstargets/pull/16)
- [X] [ generate_data in mcmc_rep.Rmd doesn't use alpha? #17](https://github.com/wlandau/jagstargets/issues/17)
