+++ inst/paper.bib
@@ -110,3 +110,17 @@
   note = {R package version 0.6-1},
   url = {https://CRAN.R-project.org/package=R2jags},
 }
+
+@article{
+  stantargets,
+  doi = {10.21105/joss.03193},
+  url = {https://doi.org/10.21105/joss.03193},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {60},
+  pages = {3193},
+  author = {William Michael Landau},
+  title = {The stantargets {R} package: a workflow framework for efficient reproducible {S}tan-powered {B}ayesian data analysis pipelines},
+  journal = {Journal of Open Source Software}
+}
