+++ inst/paper.md
@@ -28,8 +28,8 @@ Bayesian statisticians regularly experiment with models to refine, compare, and 
 
 # Summary
 
-The [`jagstargets`](https://github.com/wlandau/jagstargets) R package [@jagstargets] is a pipeline toolkit for Bayesian data analysis with JAGS. It automatically detects the dependency relationships among the results, models, and datasets, it watches these computational steps for changes, and it deploys the work with optional distributed computing while skipping the tasks that are already synchronized with their upstream dependencies. This automated selective skipping reduces the runtime of successive experiments, and when the entire workflow is up to date, the user has tangible evidence of reproducibility.
+The [`jagstargets`](https://docs.ropensci.org/jagstargets/) R package [@jagstargets] is a pipeline toolkit for Bayesian data analysis with JAGS. It automatically detects the dependency relationships among the results, models, and datasets, it watches these computational steps for changes, and it deploys the work with optional distributed computing while skipping the tasks that are already synchronized with their upstream dependencies. This automated selective skipping reduces the runtime of successive experiments, and when the entire workflow is up to date, the user has tangible evidence of reproducibility.
 
-The [`jagstargets`](https://github.com/wlandau/jagstargets) package is a bridge between packages [`R2jags`](https://github.com/suyusung/R2jags) [@R2jags] and [`targets`](https://github.com/wlandau/targets) [@targets]. [`targets`](https://github.com/wlandau/targets) is a general-purpose pipeline toolkit for reproducible research and high-performance computing, and it is not specific to Bayesian data analysis. [`jagstargets`](https://github.com/wlandau/jagstargets) leverages the existing capabilities of [`targets`](https://github.com/wlandau/targets) to more easily and concisely express computational pipelines for Bayesian statisticians, from single analyses or large-scale simulation studies. [`jagstargets`](https://github.com/wlandau/targets) is a member of the [R Targetopia](https://wlandau.github.io/targetopia) [@targetopia], an ecosystem of workflow frameworks to make [`targets`](https://github.com/wlandau/targets) more accessible to specific domains of Statistics and data science.
+The [`jagstargets`](https://docs.ropensci.org/jagstargets/) package is a bridge between packages [`R2jags`](https://github.com/suyusung/R2jags) [@R2jags] and [`targets`](https://docs.ropensci.org/targets/) [@targets]. [`targets`](https://docs.ropensci.org/targets/) is a general-purpose pipeline toolkit for reproducible research and high-performance computing, and it is not specific to Bayesian data analysis. [`jagstargets`](https://docs.ropensci.org/jagstargets/) leverages the existing capabilities of [`targets`](https://docs.ropensci.org/targets/) to more easily and concisely express computational pipelines for Bayesian statisticians, from single analyses or large-scale simulation studies. [`jagstargets`](https://docs.ropensci.org/targets/) is similar to the [`stantargets`](https://docs.ropensci.org/stantargets/) R package [@stantargets], the latter of which streamlines pipeline construction of Bayesian data analysis pipelines with Stan [@stan].
 
 # References
