+++ R/tar_jags_package.R
@@ -2,9 +2,9 @@
 #' @docType package
 #' @description Bayesian data analysis usually incurs long runtimes
 #'   and cumbersome custom code. A pipeline toolkit tailored to
-#'   Bayesians, the `jagstargets` R package leverages
+#'   Bayesian statisticians, the `jagstargets` R package leverages
 #'   `targets` and `R2jags` to ease this burden.
-#'   `jagstargets` makes it super easy to set up useful scalable
+#'   `jagstargets` makes it super easy to set up scalable
 #'   JAGS pipelines that automatically parallelize the computation
 #'   and skip expensive steps when the results are already up to date.
 #'   Minimal custom code is required, and there is no need to manually
