+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: aae5ce1adce646d180f29ede23210ef13527ac7c
 articles:
   mcmc: mcmc.html
   mcmc_rep: mcmc_rep.html
-last_built: 2021-05-08T19:45Z
+last_built: 2021-05-08T20:12Z
 
