+++ tests/testthat/test-tar_jags_example_file.R
@@ -1,6 +1,6 @@
 # targets::tar_test() runs the test code inside a temporary directory
 # to avoid accidentally writing to the user's file space.
-targets::tar_test("tar_jags_example_file()", {
+targets::tar_test("tar_jags_example_file() creates a valid JAGS model file.", {
   tmp <- tempfile()
   expect_false(file.exists(tmp))
   tar_jags_example_file(path = tmp)
