+++ _pkgdown.yml
@@ -39,5 +39,5 @@ articles:
     desc: >
       MCMC for Bayesian models
     contents:
-      - mcmc
-      - mcmc_rep
+      - introduction
+      - simulation
