+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 2.7.3
 pkgdown: 1.6.1.9001
 pkgdown_sha: aae5ce1adce646d180f29ede23210ef13527ac7c
 articles:
-  mcmc: mcmc.html
-  mcmc_rep: mcmc_rep.html
-last_built: 2021-05-08T20:28Z
+  introduction: introduction.html
+  simulation: simulation.html
+last_built: 2021-05-08T20:39Z
 
