### 2021-05-09 0ef2c23
Reword intro vignette setup
