+++ vignettes/introduction.Rmd
@@ -91,10 +91,16 @@ list(
 )
 ```
 
-[`tar_jags()`](https://wlandau.github.io/jagstargets/reference/tar_jags.html) only *defines* the pipeline. It does not actually run JAGS, it declares the targets that will eventually run JAGS. The specific targets are as follows.
+[`tar_jags()`](https://wlandau.github.io/jagstargets/reference/tar_jags.html) only *defines* the pipeline. It does not actually run JAGS, it declares the targets that will eventually run JAGS. The specific targets are as follows. Run `tar_manifest()` to show specific details about the targets declared.
+
+```{r}
+tar_manifest()
+```
+
+Each target is responsible for a piece of the workflow.
 
 * `example_file_x`: Reproducibly track changes to the jags model file.
-* `example_data`: Run the code you supplied to `data` and return a dataset compatible with jags.
+* `example_data`: Run the code you supplied to the `data` argument of `tar_jags()` and return a dataset compatible with JAGS.
 * `example_mcmc_x`: Run the MCMC and return an object of class `rjags` from [`R2jags`](https://github.com/suyusung/R2jags).
 * `example_draws_x`: Return a friendly `tibble` of the posterior draws from `example`.
 * `example_summaries_x`: Return a friendly `tibble` of the posterior summaries from `example`. Uses [`posterior::summarize_draws()`](https://mc-stan.org/posterior/reference/draws_summary.html)
@@ -102,13 +108,7 @@ list(
 
 The suffix `_x` comes from the base name of the model file, in this case `x.jags`. If you supply multiple model files to the `jags_files` argument, all the models share the same dataset, and the suffixes distinguish among the various targets.
 
-Run `tar_manifest()` to show specific details about the targets.
-
-```{r}
-tar_manifest()
-```
-
-[`targets`](https://docs.ropensci.org/targets/) produces a graph to show the dependency relationships among the targets. Below, the MCMC depends on the model file and the data, and the draws, summary, and diagnostics depend on the MCMC.
+The targets depend on one another: for example, `example_mcmc_x` takes `example_data` as input. [`targets`](https://docs.ropensci.org/targets/) can visualize the dependency relationships in a dependency graph, which is helpful for understanding the pipeline and troubleshooting issues.
 
 ```{r, output = FALSE, message = FALSE}
 tar_visnetwork(targets_only = TRUE)
