+++ vignettes/simulation.Rmd
@@ -29,7 +29,7 @@ library(jagstargets)
 
 The [introductory vignette](https://wlandau.github.io/jagstargets/articles/introduction.html) vignette caters to Bayesian data analysis workflows with few datasets to analyze. However, it is sometimes desirable to run one or more Bayesian models repeatedly across many simulated datasets. Examples:
 
-1. Validate the implementation of a Bayesian model using simulation.
+1. Validate the implementation of a Bayesian model, using simulation to determine how reliably the model estimates the parameters under known data-generating scenarios.
 2. Simulate a randomized controlled experiment to explore frequentist properties such as power and Type I error.
 
 This vignette focuses on (1). The goal of this particular example to simulate multiple datasets from the model below, analyze each dataset, and assess how often the estimated posterior intervals cover the true parameters from the prior predictive simulations. The quantile method by @cook2006 generalizes this concept, and simulation-based calibration [@talts2020] generalizes further. The interval-based technique featured in this vignette is not as robust as SBC, but it may be more expedient for large models because it does not require visual inspection of multiple histograms.
@@ -37,10 +37,11 @@ This vignette focuses on (1). The goal of this particular example to simulate mu
 ```{r}
 lines <- "model {
   for (i in 1:n) {
-    y[i] ~ dnorm(alpha + x[i] * beta[i], 1)
+    y[i] ~ dnorm(beta[1] + x[i] * beta[2], 1)
+  }
+  for (i in 1:2) {
     beta[i] ~ dnorm(0, 1)
   }
-  alpha ~ dnorm(0, 1)
 }"
 writeLines(lines, "model.jags")
 ```
@@ -55,11 +56,10 @@ tar_script({
   options(crayon.enabled = FALSE)
   tar_option_set(memory = "transient", garbage_collection = TRUE)
   generate_data <- function (n = 10L) {
-    alpha <- stats::rnorm(n = 1, mean = 0, sd = 1)
-    beta <- stats::rnorm(n = n, mean = 0, sd = 1)
+    beta <- stats::rnorm(n = 2, mean = 0, sd = 1)
     x <- seq(from = -1, to = 1, length.out = n)
-    y <- stats::rnorm(n, alpha + x * beta, 1)
-    .join_data <- list(alpha = alpha, beta = beta)
+    y <- stats::rnorm(n, beta[1] + x * beta[2], 1)
+    .join_data <- list(beta = beta)
     list(n = n, x = x, y = y, .join_data = .join_data)
   }
   list(
@@ -67,12 +67,12 @@ tar_script({
       model,
       "model.jags",
       data = generate_data(),
-      parameters.to.save = c("alpha", "beta"),
+      parameters.to.save = "beta",
       batches = 5, # Number of branch targets.
-      reps = 2, # Number of model reps per branch target.
+      reps = 4, # Number of model reps per branch target.
       stdout = R.utils::nullfile(),
       stderr = R.utils::nullfile(),
-      variables = c("alpha", "beta"),
+      variables = "beta",
       summaries = list(
         ~posterior::quantile2(.x, probs = c(0.025, 0.975))
       )
@@ -89,14 +89,13 @@ options(crayon.enabled = FALSE)
 tar_option_set(memory = "transient", garbage_collection = TRUE)
 
 generate_data <- function (n = 10L) {
-  alpha <- stats::rnorm(n = 1, mean = 0, sd = 1)
-  beta <- stats::rnorm(n = n, mean = 0, sd = 1)
[…]
