### 2021-05-09 586f75e
Distinguish among batches, branches, reps, and iterations
