+++ man/tar_jags_example_data.Rd
@@ -39,6 +39,16 @@ element so the true value of \code{beta} from the simulation
 is automatically appended to the \code{beta} rows of the
 summary output.
 }
+\details{
+The \code{tar_jags_example_data()} function draws a JAGS
+dataset from the prior predictive distribution of the
+model from \code{\link[=tar_jags_example_file]{tar_jags_example_file()}}. First, the
+regression coefficient \code{beta} is drawn from its standard
+normal prior, and the covariate \code{x} is computed.
+Then, conditional on the \code{beta} draws and the covariate,
+the response vector \code{y} is drawn from its
+Normal(\code{x * beta}, 1) likelihood.
+}
 \examples{
 tar_jags_example_data()
 }
