### 2021-05-09 #425 @wlandau
Thanks for your review, @dill. Your feedback helped clean up the documentation and make it more accessible to new users.

> but (2021/04/29) targets had been removed from CRAN, so this failed with error:

`targets` is back up on CRAN, and I resubmitted `tarchetypes` on May 4. If all goes well, it may be accepted next week.

> covr::package_coverage(type="all") reports: 99.57% coverage

That's odd, https://app.codecov.io/gh/wlandau/jagstargets and my local machines both report 100% coverage. Do you happen to know if any tests were skipped due to missing optional packages?

> The first two paragraphs give a summary of the package but they feel like they are written for people who already know what targets does (e.g., "factories", "targets", "pipelines" are all jargon to begin with; though I appreciated the links to the blog post, I can see this turning people off right away). Perhaps there could be some adaptation so that someone who is a JAGS user can immediately understand how the package will save them time and headaches. Maybe even a "for JAGS users" and a "for people familiar with targets" section might help? (FWIW, I found the text in ?jagstargets-package to be a bit more transparent.)

In https://github.com/wlandau/jagstargets/commit/b6d12e3d75361a24feacc74ec6988ee994741936, I rewrote the introductory paragraph and moved the technical details to a later section of the README.

> Tests didn't use context() to provide extra information on what exactly was being tested each time. Filenames were semi-informative, but might be better to have more info there for contributors. This might be more useful if tests are expanded further.

`testthat::context()` is superseded, and I am trying to follow the updated recommendation to use file names as context.

> Not much in the way of comments in the test files. Some are self-explanatory but the large tests are harder to follow.

https://github.com/wlandau/jagstargets/commit/6528f1475c1a7fe3151654905e7c65625b050ffa inserts more inline prose to explain the intent of the tests.

> Could this be renamed to "Getting started with..." or "An introduction to..." to make it clear this is where to start?

Sure, implemented in https://github.com/wlandau/jagstargets/commit/885cce7f36a29a7203afba97717b41b8cda21780.

> It might be nice to have a bit of table setting at the start of this vignette to talk about the kinds of outputs we want, even showing how an existing JAGS workflow works and how that maps into jagstargets (and then you can say how similar this is to the DAG generated by tar_visnetwork? In order to get people on board I think it's necessary to show clear mapping between their current workflow and "doing something fancy" and a bit of intro here might help that happen.

I am not sure I fully understand, but https://github.com/wlandau/jagstargets/commit/63ad892b620ab706188fe855cd5c2ff1de1b8023 does work a little harder to put the `targets` pipeline in the context of an ordinary JAGS workflow. Please let me know if this different from what you have in mind.

> I found the list of targets a little confusing initially (around here). I wasn't sure where I was supposed to be looking for those objects and what was going on. Perhaps running tar_manifest() first, showing the effect of the tar_jags call, then explaining what these entries mean might make that easier to understand?

https://github.com/wlandau/jagstargets/commit/0ef2c23a3c2db1ebf968738d1f98618d443179f4 rewords that section to try to be more helpful.

> Does the tar_jags call have to be wrapped in list() in that chunk (if so maybe explain why).

`list()` is not required in this specific instance, but it underscores that `_targets.R` should end with a list of targets in the general case. https://github.com/wlandau/jagstargets/commit/e371d0f24eebc1e149d7d9c80c4b712207511fc3 adds a comment to clarify this. (Returning a single target is also acceptable but not very common.)

> The last part of the vignette starts talking about a second model y.jags though we don't define that. Is it worth make just a quick note here that tar_visnetwork doesn't check that the model files exist?

`tar_jags()` (and thus `tar_visnetwork()`) does error out if the JAGS file does not exist. (The original vignette wrote the model in a hidden code chunk.) https://github.com/wlandau/jagstargets/commit/cbe845e9dc8ca6415c14b238abb66b95e3fc67f1 exposes the new model to the reader.

> Might be worth having something up top saying to start with the other vignette first?

Added a mention of the intro vignette.

> There could be a little more chat at the start of this vignette explaining that you want to find out about how well model.jags estimates alpha and beta[].

https://github.com/wlandau/jagstargets/commit/c5af606cdd34802b34767ec729b666b4b3ea066c adds a comment near the beginning to clarify a bit about what I mean by "validation".

> This is a bit of an odd model (in my opinion at least!) in that you have fixed the covariate values and have beta be a random effect (you generate n random slopes), rather than a fixed beta with random data each time. I wonder if a classic linear model situation (where data vary and there is only one beta) would be a clearer situation for people to think about? It also has the advantage of producing much less output, as well as having the same DGP as the other vignette.

https://github.com/wlandau/jagstargets/commit/c5af606cdd34802b34767ec729b666b4b3ea066c changes the model to have one slope and one intercept. My intention with the different model was to demonstrate that `jagstargets` can handle non-scalar parameters when it comes to simulation studies that look at coverage in posterior intervals.

> With my stats hat on, I'd say write out the equations for the model/DGP, so it's super-clear what's happening here.

Added in https://github.com/wlandau/jagstargets/commit/057c47cc68199992fadd076660f877c6f731cd5b.

> There is some setup in the chunks here that isn't explained (e.g., options(crayon.enabled = FALSE), tar_option_set(memory = "transient", garbage_collection = TRUE) here that is worth explaining to novice users like myself. Perhaps this is only necessary for this vignette format, but it would be good to know that if so.

`options(crayon.enabled = FALSE)` is internal to the vignette. It makes sure printed output is rendered properly in the R Markdown document, and it does not appear in the rendered HTML document. https://github.com/wlandau/jagstargets/commit/4ddaf4444839d01ae0b857cafb65736314fafb09 adds a comment to indicate that the other options help `targets` behave more efficiently with computer memory. Those options are explained in more detail in `?targets::tar_option_set`.

> I found the _targets.R definition hard to follow, particularly the arguments to tar_jags_rep_summary() were hard to relate to what is going on...
> The terminology could be explained here at the same time: "branches", "replications", "batches" and "iterations" all appear almost at once. I think I've figured out what they mean by the end, but it would be nice to have that clear as those terms are and why we use them here. This could even be woven into the first 1-2 intro paragraphs?

Hopefully https://github.com/wlandau/jagstargets/commit/586f75e73d87ccfc513216df89edccae243d3baa clears that up.

> The names of the nodes in these graphs are hard-going (this was less obvious above when the model was simpler). model_model vs. model vs. model_lines_model: these make sense once you know what you're doing I'm sure but I found them difficult to keep track of (see comment in "Package code" below). Worth pointing people to the documentation in ?tar_jags to explain the naming convention at this point.

From the [documentation of `tar_jags()`](https://github.com/wlandau/jagstargets/blob/586f75e73d87ccfc513216df89edccae243d3baa/R/tar_jags.R#L11-L15)

```r
#'   The target names use the `name` argument as a prefix, and the individual
#'   elements of `jags_files` appear in the suffixes where applicable.
#'   As an example, the specific target objects returned by
#'   `tar_jags(name = x, jags_files = "y.jags", ...)` returns a list
#'   of `targets::tar_target()` objects:
```

Still takes getting used to though, I know.

> Is there any useful information that can be gleaned from the .rep column? It looks like a hash, can it tell us anything re: reproducibility or is it just for internal bookkeeping?

Yes, it is just a hash to distinguish among individual simulation replications. [Dynamic branching in `targets`](https://books.ropensci.org/targets/dynamic.html) automatically mashes all the reps together in the same data frame, and having a rep ID helps if the user wants to do custom post-processing within reps, e.g. `tar_read(model) %>% group_by(.rep) %>% group_modify(~derive_special_parameter(.x)) %>% ...`.

> I tried to construct a comparison between mgcv::gam and mgcv::jagam using jagstargets. You can find the resulting code here. Mostly I found this fairly intuitive once I got my head around the above. tar_visnetwork() was essential to debugging. Feel no obligation to look at this but it might be useful to see what weird things people do when working from scratch and seeing where some bottlenecks are.

Awesome! Thanks for sharing.

> If jagstargets-package is a landing page, perhaps it could contain a link back to the pkgdown site, and perhaps a @seealso to the most useful starting manpage(s) (tar_jags?)

Added in https://github.com/wlandau/jagstargets/commit/240d5cdd72744f32baeba32827cd29dbc60c031e.

> tar_jags_example_data.R should explain the data generating process, as well as what the components of the returned object.

https://github.com/wlandau/jagstargets/commit/3a8fb6775e1a1e635846f827664af152b007bef1 adds a `@details` section to explain the data-generating process (drawing from the prior predictive distribution). `tar_jags_example_data()` has a `@format` section to explain the elements of the returned JAGS data list.

> Naming scheme for targets: are there more descriptive names for the x_file_y and x_lines_y? I don't have stellar suggestions on replacements but "file" and "lines" feel very generic and like they might get confused in a large workflow? The other labels seem very clear though.
> Following-on from this, it is possible to combine the x_file_y and x_lines_y targets? I see in the stantargets vignette there doesn't seem to be a "compile" step, just the x_file_y then x_mcmc_y -- is there something that is different here that means we need both steps to be separate?

All this has to do with how `jagstargets` and `stantargets` handle high-performance computing. Some users just run models on their local laptops, while others distribute multiple models on computing clusters. On clusters, some users configure jobs to run at the root directory of the project, while others use node-specific storage. And eventually (hopefully) the parallel backends of `targets` will be able to submit jobs to the cloud, and those environments will not have direct access to the original JAGS model files the users start with.

`jagstargets` is the simpler case. All we need to do is:

1. Hash the upstream JAGS model file and watch it for changes.
2. Make sure parallel workers have access to the JAGS model even if they cannot access the original model file.

I solve (2) by reading the actual lines of text of the JAGS model file in a separate target. A downstream MCMC target writes these lines to a new temporary JAGS model file and runs the temporary file instead of the user's original file. Because of the way [file tracking](https://books.ropensci.org/targets/files.html#external-input-files) works in `targets`, (1) and (2) need to be separate targets and cannot be combined into one.

`stantargets` is more complicated because Stan models require compilation. On systems where parallel workers do not share storage, we need the "lines" trick so the workers have access to the model, and that means each worker needs to compile its own copy of the model. But compilation takes a lot of time, so on systems that let parallel workers share storage efficiently, `stantargets` should pre-compile each model once ahead of time and then let the workers run in parallel on the pre-compiled models. The `compile` argument allows the user to choose when to compile models, and this argument changes which targets are created and what they do.

I hope this clarifies the roles of the "file" and "lines" targets. They do need to be separate, and I do think these names accurately describe what the targets do, even if the specifics are esoteric for most users. Happy to consider specific suggestions for alternative names if you have any.
