### 2021-05-12 #425 @wlandau
>coverage: I looked at the report and it says tar_jags_df.R lines 68:70 are not covered (relating to tar_jags_df_summary()), so that doesn't look like an optional package issue?

Thanks, that was a big hint. Hopefully https://github.com/wlandau/jagstargets/commit/58621cbc092a3016dac07784ebbb4fb37797031c has better coverage.

> I guess I'm just thinking about a very specific use case where I might compare similar models fitted in JAGS vs Stan and how I can tell which bit of the model is which

Yeah, `jagstargets` and `stantargets` remove file extensions when they compute target names. So users who call both in the same pipeline would need to disambiguate the model files in other ways. Not ideal for that specific use case, but `targets` lets you know pretty quickly when there are name conflicts. On my end, changing the name policy is always a risk because it invalidates everyone's workflows so they have to rerun those targets.
