+++ pkgdown.yml
@@ -1,8 +1,8 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9001
-pkgdown_sha: aae5ce1adce646d180f29ede23210ef13527ac7c
+pkgdown_sha: 71849077a7232b20d0f76eb7203e558b437e0dd7
 articles:
   introduction: introduction.html
   simulation: simulation.html
-last_built: 2021-05-09T15:50Z
+last_built: 2021-05-12T05:18Z
 
