### 2021-05-12 #425 @dill
Re-running `covr::package_coverage(type="all")` on  wlandau/jagstargets@58621cb gives 100% coverage 🎉

That's helpful extra info on name conflicts, thanks for that!
