+++ R/tar_jags_rep_dic.R
@@ -84,6 +84,8 @@ tar_jags_rep_dic <- function(
   tidy_eval = targets::tar_option_get("tidy_eval"),
   packages = targets::tar_option_get("packages"),
   library = targets::tar_option_get("library"),
+  format = "qs",
+  format_df = "fst_tbl",
   error = targets::tar_option_get("error"),
   memory = targets::tar_option_get("memory"),
   garbage_collection = targets::tar_option_get("garbage_collection"),
@@ -120,6 +122,8 @@ tar_jags_rep_dic <- function(
     tidy_eval = tidy_eval,
     packages = packages,
     library = library,
+    format = format,
+    format_df = format_df,
     error = error,
     memory = memory,
     garbage_collection = garbage_collection,
