+++ R/tar_jags_rep_summary.R
@@ -87,6 +87,8 @@ tar_jags_rep_summary <- function(
   tidy_eval = targets::tar_option_get("tidy_eval"),
   packages = targets::tar_option_get("packages"),
   library = targets::tar_option_get("library"),
+  format = "qs",
+  format_df = "fst_tbl",
   error = targets::tar_option_get("error"),
   memory = "transient",
   garbage_collection = targets::tar_option_get("garbage_collection"),
@@ -126,6 +128,8 @@ tar_jags_rep_summary <- function(
     tidy_eval = tidy_eval,
     packages = packages,
     library = library,
+    format = format,
+    format_df = format_df,
     error = error,
     memory = memory,
     garbage_collection = garbage_collection,
