+++ R/tar_jags.R
@@ -69,6 +69,16 @@
 #'   data frame of posterior summary statistics and convergence diagnostics.
 #' @param dic Logical, whether to create a target with deviance
 #'   information criterion (DIC) results.
+#' @param format Character of length 1, storage format of the non-data-frame
+#'   targets such as the JAGS data and any JAGS fit objects.
+#'   Please choose an all=purpose
+#'   format such as `"qs"` or `"aws_qs"` rather than a file format like
+#'   `"file"` or a data frame format like `"parquet"`. For more on storage
+#'   formats, see the help file of `targets::tar_target()`.
+#' @param format_df Character of length 1, storage format of the data frame
+#'   targets such as posterior draws. We recommend efficient data frame formats
+#'   such as `"feather"` or `"aws_parquet"`. For more on storage formats,
+#'   see the help file of `targets::tar_target()`.
 #' @examples
 #' targets::tar_dir({ # tar_dir() runs code from a temporary directory.
 #' targets::tar_script({
@@ -121,6 +131,8 @@ tar_jags <- function(
   tidy_eval = targets::tar_option_get("tidy_eval"),
   packages = targets::tar_option_get("packages"),
   library = targets::tar_option_get("library"),
+  format = "qs",
+  format_df = "fst_tbl",
   error = targets::tar_option_get("error"),
   memory = targets::tar_option_get("memory"),
   garbage_collection = targets::tar_option_get("garbage_collection"),
@@ -234,7 +246,7 @@ tar_jags <- function(
     command = command_data,
     packages = packages,
     library = library,
-    format = "qs",
+    format = format,
     error = error,
     memory = memory,
     garbage_collection = garbage_collection,
@@ -245,7 +257,7 @@ tar_jags <- function(
   target_object_mcmc <- targets::tar_target_raw(
     name = name_mcmc,
     command = command_mcmc,
-    format = "qs",
+    format = format,
     packages = character(0),
     error = error,
     memory = memory,
@@ -261,7 +273,7 @@ tar_jags <- function(
     name = name_draws,
     command = command_draws,
     packages = character(0),
-    format = "fst_tbl",
+    format = format_df,
     error = error,
     memory = memory,
     garbage_collection = garbage_collection,
@@ -273,7 +285,7 @@ tar_jags <- function(
     name = name_summary,
     command = command_summary,
     packages = character(0),
-    format = "fst_tbl",
+    format = format_df,
     error = error,
     memory = memory,
     garbage_collection = garbage_collection,
@@ -285,7 +297,7 @@ tar_jags <- function(
     name = name_dic,
     command = command_dic,
     packages = character(0),
-    format = "fst_tbl",
+    format = format_df,
     error = error,
     memory = memory,
     garbage_collection = garbage_collection,
