+++ pkgdown.yml
@@ -1,8 +1,8 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9001
-pkgdown_sha: 71849077a7232b20d0f76eb7203e558b437e0dd7
+pkgdown_sha: 5bd377e7cb6d268565018b73e4bb99d2788182ed
 articles:
   introduction: introduction.html
   simulation: simulation.html
-last_built: 2021-05-12T05:18Z
+last_built: 2021-05-22T13:30Z
 
