### 2021-06-17 a85d7b0
Update intro vignette to arg name change
