+++ pkgdown.yml
@@ -1,8 +1,8 @@
 pandoc: 2.7.3
 pkgdown: 1.6.1.9001
-pkgdown_sha: 5bd377e7cb6d268565018b73e4bb99d2788182ed
+pkgdown_sha: 9ce6af43b55cd21e9563526e12a12d1243657bf6
 articles:
   introduction: introduction.html
   simulation: simulation.html
-last_built: 2021-05-22T13:30Z
+last_built: 2021-06-17T03:58Z
 
