+++ R/utils_assert.R
@@ -2,6 +2,6 @@ assert_jags_file <- function(jags_file) {
   targets::tar_assert_chr(jags_file)
   targets::tar_assert_path(jags_file)
   if (dir.exists(jags_file)) {
-    throw_validate("jags_file must not be a directory.")
+    targets::tar_throw_validate("jags_file must not be a directory.")
   }
 }
