+++ NAMESPACE
@@ -47,6 +47,7 @@ importFrom(targets,tar_script)
 importFrom(targets,tar_target)
 importFrom(targets,tar_target_raw)
 importFrom(targets,tar_test)
+importFrom(targets,tar_throw_validate)
 importFrom(targets,tar_tidy_eval)
 importFrom(tibble,as_tibble)
 importFrom(tibble,tibble)
