+++ R/tar_jags_package.R
@@ -23,7 +23,8 @@
 #' @importFrom rlang expr quo_squash
 #' @importFrom stats rnorm runif
 #' @importFrom targets tar_assert_chr tar_assert_in
-#'   tar_assert_nonempty tar_assert_not_in tar_assert_nzchar
+#'   tar_assert_nonempty tar_assert_not_dir
+#'   tar_assert_not_in tar_assert_nzchar
 #'   tar_assert_path tar_assert_scalar tar_assert_unique
 #'   tar_deparse_language tar_deparse_safe
 #'   tar_cue tar_dir tar_load tar_option_get tar_path
