+++ R/utils_assert.R
@@ -1,7 +1,5 @@
 assert_jags_file <- function(jags_file) {
   targets::tar_assert_chr(jags_file)
   targets::tar_assert_path(jags_file)
-  if (dir.exists(jags_file)) {
-    targets::tar_throw_validate("jags_file must not be a directory.")
-  }
+  targets::tar_assert_not_dir(jags_file)
 }
