+++ NAMESPACE
@@ -30,6 +30,7 @@ importFrom(tarchetypes,tar_map)
 importFrom(targets,tar_assert_chr)
 importFrom(targets,tar_assert_in)
 importFrom(targets,tar_assert_nonempty)
+importFrom(targets,tar_assert_not_dir)
 importFrom(targets,tar_assert_not_in)
 importFrom(targets,tar_assert_nzchar)
 importFrom(targets,tar_assert_path)
