+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: 51268b0d9cb0588d7a6404a2b8ef4ee19c435e89
 articles:
   introduction: introduction.html
   simulation: simulation.html
-last_built: 2021-06-21T20:42Z
+last_built: 2021-06-21T21:27Z
 
