### 2021-07-14 #425 @ernestguevarra
On running the workflow in the worked example in the vignette entitled "Bayesian simulation pipelines with jagstargets" that adds a target called *coverage*, I see the following errors:

```R
Error: callr subprocess failed: could not find function "%>%"
```

```R
Error: callr subprocess failed: could not find function "summarize"
```

In the `_targets.R` file, I added the following call:

to be able to get the updated workflow running.

When I reviewed `simulation.Rmd` which is the Rmd for this vignette, I noted that in the setup R code chunk, calls to load the following packages were made:

```R
library(R2jags)
library(targets)
library(jagstargets)
```

I am unsure if this is correct thinking on my part but I am guessing because `{dplyr}` is loaded in the setup code chunk of the vignette that when building the vignette, the specific example is not erroring but when I run the example outside of the Rmd that the workflow errors with the message above? If so, would it be appropriate to add a call to load the package `{dplyr}` in the `_targets.R` file example as follows:

```R
# _targets.R
library(targets)
library(jagstargets)

generate_data <- function (n = 10L) {
  beta <- stats::rnorm(n = 2, mean = 0, sd = 1)
  x <- seq(from = -1, to = 1, length.out = n)
  y <- stats::rnorm(n, beta[1] + x * beta[2], 1)
  # Elements of .join_data get joined on to the .join_data column
  # in the summary output next to the model parameters
  # with the same names.
  .join_data <- list(beta = beta)
  list(n = n, x = x, y = y, .join_data = .join_data)
}

list(
  tar_jags_rep_summary(
    model,
    "model.jags",
    data = generate_data(),
    parameters.to.save = "beta",
    batches = 5, # Number of branch targets.
    reps = 4, # Number of model reps per branch target.
    variables = "beta",
    summaries = list(
      ~posterior::quantile2(.x, probs = c(0.025, 0.975))
    )
  ),
  tar_target(
    coverage,
    model %>%
      group_by(variable) %>%
      summarize(
        coverage = mean(q2.5 < .join_data & .join_data < q97.5),
        .groups = "drop"
      )
  )
)
```

I am a newbie with using targets and these are common errors that I encounter also and my usual approach will be to make a call within the targets workflow to load the package that is causing the error. Whether this is the correct approach, I am not sure but I am flagging it here as others who would follow the example may face the same errors I saw/faced.

- [X] **Examples** (that run successfully locally) for all exported functions

On install of `{jagstargets}` using `devtools::install(dependencies = TRUE)` from within the project, I see the following error:

```
Error: object ‘tar_assert_chr’ is not exported by 'namespace:targets'

In i.p(...) :
  installation of package ‘/tmp/RtmpsW4y3O/file615c3eabc513/tarchetypes_0.2.1.9000.tar.gz’ had non-zero exit status
```

But `{tarchetypes}` package was installed. I think this is more an issue with `{tarchetypes}` maybe?

On `covr::package_coverage()`, I see the following result:

```R
jagstargets Coverage: 14.46%
R/tar_jags_rep.R: 0.00%
R/utils_language.R: 0.00%
R/tar_jags.R: 1.72%
R/tar_jags_rep_summary.R: 2.44%
R/tar_jags_rep_dic.R: 2.63%
R/tar_jags_rep_draws.R: 2.63%
R/tar_jags_df.R: 100.00%
R/tar_jags_example_data.R: 100.00%
R/tar_jags_example_file.R: 100.00%
R/utils_assert.R: 100.00%
R/utils_data.R: 100.00%
R/utils_jags.R: 100.00%
R/utils_logic.R: 100.00%
```

However, `codecov` results show 100% coverage across the same functions. I am unsure why this is the case but maybe from new edits to the functions since last code coverage check? I may have contributed to this due to my tardiness in the package review so my sincerest apologies for this.

The package is consistent with rOpenSci packaging guidelines and also equally importantly with the `{targets}` ecosystem within which it belongs. Given that the `{targets}` package is also rOpenSci reviewed has helped ensure the thematic and packaging coherence of `{jagstargets}`. The package name, branding, tone and construct of README and vignettes all are coherent as a standalone package and also consistent within the package ecosystem it inhabits. As a newbie and recent user of `{targets}`, I found the `{targets}` story continue well into `{jagstargets}` particularly the function naming and documentation, the README/website, and the vignettes.

---

Firstly, I would like to apologise for the extreme delay in reviewing this package. I have been severely disorganised with my work and this has impacted on my ability to get to this review promptly.

I thank @wlandau for his work on the `{jagstargets}` which, from my review, has built on the strength of the `{targets}` package and facilitates its easy application for Bayesian data analysis using JAGS. I approach this review from the perspective of someone who hasn't formally used JAGS and someone who has only recently learned and currently using `{targets}`. The author's note in the README of requiring basic knowledge of `{targets}` and familiarity with Bayesian statistics and JAGS fit me as a new user and because of two really well worked out vignettes on how to get started and on simulation, I was able to go through a quick and easy introduction and orientation to the package and gives me enough foundation to be able to apply the package to appropriate tasks and related work in the future. It also definitely helps that `{targets}` is well documented with lots of really good examples and has an accompanying manual/handbook reference that I was able to use before to learn how to use it. With the `{jagstargets}` package being coherently written and documented in relation to the `{targets}` package, anyone who has already invested time in learning `{targets}` will find the `{jagstargets}` extremely easy to learn how to use and immediately handy in implementing a `{targets}` workflow for JAGS easily and quickly.

I have made a few comments above in the appropriate review topics on some minor issues I encountered when installing the development version of the package, on following one example in the simulation vignette, on code coverage, and on Good Practice check.

I would also like to mention that there were few and very minor edits that can be made in the vignettes (missing words, punctuation) that don't impact on readability and comprehension of the material but the author might want to fully correct/edit. If so, I would be happy to make those edits into a pull request.

Thank you again to @wlandau for your great work on this package and for the whole `{targets}` package ecosystem that you have and are continuing to create.
