### 2021-07-14 #425 @wlandau
Thank you for your review, @ernestguevarra.

> On running the workflow in the worked example in the vignette entitled "Bayesian simulation pipelines with jagstargets" that adds a target called coverage, I see the following errors...

https://github.com/wlandau/jagstargets/commit/30e4ef60214bcdcb8558e07f1011e4b117305f49 repairs the `_targets.R` files for that example. The problem has to do with some code duplication that I unfortunately found necessary at the time to embed the example in an R Markdown vignette: there is [one code chunk to show `_targets.R` to the reader](https://github.com/wlandau/jagstargets/blob/30e4ef60214bcdcb8558e07f1011e4b117305f49/vignettes/simulation.Rmd#L209-L255) and [another code chunk to actually write `_targets.R`](https://github.com/wlandau/jagstargets/blob/30e4ef60214bcdcb8558e07f1011e4b117305f49/vignettes/simulation.Rmd#L160-L207).

(Knowing what I know now after implementing [Target Markdown](https://books.ropensci.org/targets/markdown.html), I think I could hack a [custom `knitr` engine](https://bookdown.org/yihui/rmarkdown/language-engines.html) that writes a script file instead of evaluating the code, and this should help avoid duplicated code in new documentation I create from scratch for projects in the future.)

> On install of {jagstargets} using devtools::install(dependencies = TRUE) from within the project, I see the following error

This is because `jagstargets` currently relies on some functions exposed in development `targets` that are not yet on CRAN (see `Remotes: ropensci/targets` in the `jagstargets` `DESCRIPTION` file). (I plan to update `targets` on CRAN on July 21.) If you run `remotes::install_github("ropensci/targets")` first, you should be able to install `jagsttargets`.

> On covr::package_coverage(), I see the following result

Some tests use `testthat::skip_on_cran()` in order to reduce check time and other CRAN-related issues. If you set the `NOT_CRAN` environment variable to `"true"`, these tests should run and coverage should show as 100%. (See `usethis::edit_r_environ()`.) When I commented out `NOT_CRAN=true` in my own `~/.Renviron` file, I saw the exact coverage pattern you posted.

> fix this R CMD check WARNING: LaTeX errors when creating PDF version. This typically indicates Rd
    problems. LaTeX errors found:

Often this particular error has something to do with the local LaTeX installation, e.g. https://stackoverflow.com/questions/10819959/diagnosing-r-package-build-warning-latex-errors-when-creating-pdf-version. I was unable to reproduce it locally.

> I would also like to mention that there were few and very minor edits that can be made in the vignettes (missing words, punctuation) that don't impact on readability and comprehension of the material but the author might want to fully correct/edit. If so, I would be happy to make those edits into a pull request.

Happy to accept pull requests about spelling and grammar, or correct any specific spelling/grammar issues you identify.

Please let me know if there is something I missed in my response.
