+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ce9781a15c7ea07df9fb17a11295ba4abec0b54b
 articles:
   introduction: introduction.html
   simulation: simulation.html
-last_built: 2021-07-14T13:12Z
+last_built: 2021-07-14T13:42Z
 
