+++ DESCRIPTION
@@ -42,7 +42,7 @@ Imports:
   rlang (>= 0.4.10),
   rjags (>= 4.10),
   stats,
-  targets (>= 0.5.0.9001),
+  targets (>= 0.6.0),
   tarchetypes (>= 0.0.1),
   tibble (>= 3.0.1),
   tools,
