+++ DESCRIPTION
@@ -35,7 +35,7 @@ Imports:
   coda (>= 0.19.4),
   digest (>= 0.6.25),
   fst (>= 0.9.2),
-  posterior (>= 0.1.2),
+  posterior (>= 1.0.1),
   purrr (>= 0.3.4),
   qs (>= 0.23.2),
   R2jags (>= 0.6.1),
@@ -57,10 +57,6 @@ Suggests:
   testthat (>= 3.0.0),
   tidyr (>= 1.1.2),
   visNetwork (>= 2.0.9)
-Remotes:
-  ropensci/targets,
-  ropensci/tarchetypes,
-  stan-dev/posterior
 SystemRequirements: JAGS 4.x.y (http://mcmc-jags.sourceforge.net)
 Encoding: UTF-8
 Language: en-US
