### 2021-08-03 e27d1ef
Add @dill as reviewer as requested at https://github.com/ropensci/software-review/issues/425#issuecomment-833703928
