+++ DESCRIPTION
@@ -26,6 +26,12 @@ Authors@R: c(
     comment = c(ORCID = "0000-0003-1878-3253")
   ),
   person(
+    given = c("David", "Lawrence"),
+    family = "Miller",
+    role = "rev",
+    email = "dave@ninepointeightone.net"
+  ),
+  person(
     family = "Eli Lilly and Company",
     role = "cph"
   ))
