### 2021-08-04 #425 @noamross
Hi @wlandau, I tend to expect a second response with changes incorporating both reviewers feedback, but I realize your workflow incorporated both responses and all edits.  That said, @ernestguevarra and @dill, do @wlandau's changes address your comments?
