### 2021-09-28 #425 @wlandau
Looks like the `jagstargets` site is up, and I think I completed the other to-dos in this thread. Remaining to-dos on my end are [JOSS](https://github.com/ropensci/jagstargets/blob/main/inst/paper.md) and then CRAN. In the meantime, @ernestguevarra, feel free to snag that "rev" role whenever you want.
